- send_wal and WalReader are now async
- notify_wal_senders is replaced with watch::Receiver<Lsn> for commit_lsn updates
- WalReader has a fallback to download segment from S3, if not available locally
- test_s3_wal_replay checks that WAL can be replayed after offloaded

Closes #1920