@petuhovskiy 2022-06-16
It looks like some debug tests are just very slow in CI. Not sure if this is connected to send_wal rewrite or not.