@petuhovskiy 2022-06-17
Duration of `test_parallel_copy` running locally:
- on main
  * debug 116s
  * release 54s
- on this branch
  * debug 132s
  * release 61s

So, this patch is definitely affecting performance and I'll try to fix this before merging. Not sure what's the best performance test or metrics for that, probably will create a new test or continue to compare `test_parallel_copy` time.