@petuhovskiy 2022-06-17
The problem was easy to find, using `read_exact` instead of `read` in WalReader is faster and makes such a difference. It seems that we probably can go even faster, if we switch to async `PostgresBackend` in send_wal and will be reading from disk and sending to the network in parallel.