@petuhovskiy 2022-06-17
It's sad but it seems that WAL redo in debug mode is still very slow. It can't replay ~48MB of WAL (3 segments) in 1 minute, which is a timeout for `wait_for_lsn`. I'll try to trigger walreceiver on pageserver somehow before calling basebackup, so we will not a have a limitation of `wait_for_lsn` timeout. If you know how to do that easily, you can write me a DM.